[오혜인] 1) addNotice.jsp 생성 > 공지사항 입력 폼 생성 , summerNote 적용 2) vo - addNoticeForm 생성 3) getNoticeListByPage.jsp 수정 > 매니저급 이상일때만 버튼 나오게 구현 4) getNoticeOne.jsp 생성 > 공지사항 상세보기 구현 5) vo - notice 수정 > noticePw 추가 6) vo - noticeFile 수정 > noticeFileSize - int -> long 타입 수정 7) NoticeController 수정 > loginCheck/getNoticeOne() getMapping, PostMapping 추가 8) NoticeFileMapper , xml , NoticeMapper , xml 수정 > 일부 메서드 매개변수 값 수정 9) NoticeService 수정 > getNoticeOne(), addNotice() 추가

+++ lms/src/main/java/kr/co/gdu/lms/vo/NoticeFile.java
@@ -10,6 +10,6 @@ public class NoticeFile {
 	private String noticeFileName;
 	private String noticeFileOriginName;
 	private String noticeFileType;
-	private int noticeFileSize;
+	private long noticeFileSize;
 	private String noticeFileCreateDate;
 }
